+++ src/converters/seqnToSeqJson.test.ts
@@ -9,10 +9,9 @@ import {
 } from '@nasa-jpl/aerie-ampcs';
 import type { VariableDeclaration } from '@nasa-jpl/seq-json-schema/types';
 import { readFileSync } from 'fs';
-import { seqJsonToSequence } from './from-seq-json';
-import { SeqnParser } from './seq-n';
-import { parser } from './seq-n.grammar.js';
-import { parseVariables, sequenceToSeqJson } from './to-seq-json';
+import { seqJsonToSequence } from './seqJsonToSeqn.js';
+import { SeqnParser } from '../languages/seq-n/seq-n.js';
+import { parseVariables, seqnToSeqJson } from './seqnToSeqJson.js';
 import { describe, expect, it } from 'vitest';
 
 function argArrToMap(cmdArgs: FswCommandArgument[]): FswCommandArgumentMap {
@@ -84,7 +83,7 @@ HDW_CMD`;
       id: 'test',
       metadata: {},
     };
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
     expect(actual).toEqual(expectedJson);
   });
 
@@ -127,7 +126,7 @@ ECHO "hello"
       ],
       metadata: {},
     };
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
     expect(actual).toEqual(expectedJson);
   });
 
@@ -150,7 +149,7 @@ HDW_CMD_2
       id: 'test',
       metadata: {},
     };
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
     expect(actual).toEqual(expectedJson);
   });
 
@@ -208,7 +207,7 @@ C FSW_CMD_1 0.123 -2.34 # inline description
         },
       ],
     };
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
     expect(actual).toEqual(expectedJson);
   });
 
@@ -273,7 +272,7 @@ R71 ECHO    L02STR
         },
       ],
     };
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
     expect(actual).toEqual(expectedJson);
   });
 
@@ -335,7 +334,7 @@ R10 PACKAGE_BANANA     2      [    "bundle1"    5 "bundle2" 10]
         },
       ],
     };
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
     expect(actual).toEqual(expectedJson);
   });
 
@@ -347,7 +346,7 @@ C ECHO L00STR
 C ECHO "L00STR"
 C ECHO L01STR
     `;
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
     const expectedJson = {
       id: 'test.inline',
       locals: [
@@ -425,7 +424,7 @@ C ECHO L00STR
 C ECHO "L01"
 C ECHO SIZE
     `;
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
     const expectedJson = {
       id: 'test.inline',
       locals: [
@@ -539,7 +538,7 @@ C ECHO SIZE
 
     for (const ordering of permutations) {
       const input = ordering.join('\n\n');
-      const actual = sequenceToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
+      const actual = seqnToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
       const expected = {
         id: 'test.seq',
         locals: [
@@ -581,7 +580,7 @@ C ECHO SIZE
 
     R1 ECHO "Can this handle \\"Escaped\\" quotes??" # and this "too"`;
     const id = 'escaped_quotes';
-    const actual = await sequenceToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
+    const actual = await seqnToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
     const expected = `{
   "id": "escaped_quotes",
   "metadata": {},
@@ -619,7 +618,7 @@ R00:00:01 ECHO "Can this handle \\"Escaped\\" quotes??" # and this "too"
 @MODEL "Variable" 0 "Offset"
 @MODEL "Variable \\"Escaped\\"" 0 "Offset \\" \\" \\"\\""`;
     const id = 'escaped_metadata';
-    const actual = await sequenceToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
+    const actual = await seqnToSeqJson(SeqnParser.parse(seq), seq, commandBanana, id);
     const expected = `{
   "id": "escaped_metadata",
   "metadata": {},
@@ -700,7 +699,7 @@ R123T11:55:33 @GROUND_EVENT("ground_event.name") "foo" 1 2 3
 R123T12:34:56 @ACTIVATE("act2.name") "foo" 1 2 3  # Comment text
 @ENGINE -1
     `;
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandBanana, '');
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandBanana, '');
     const expectedJson = {
       id,
       metadata: {},
@@ -835,7 +834,7 @@ A2024-123T12:34:56 @REQUEST_BEGIN("request.name") # Description Text
   "boolean": true
 }
     `;
-    const actual = sequenceToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
+    const actual = seqnToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
     const expected = {
       id: 'id',
       metadata: {},
@@ -911,7 +910,7 @@ G+03:00:00 "GroundEpochName" @REQUEST_BEGIN("request2.name")
 @REQUEST_END
 @METADATA "req_0_meta_name" "req_0_meta_value"
     `;
-    const actual = sequenceToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
+    const actual = seqnToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
     const expected = {
       id: 'id',
       metadata: {},
@@ -1000,7 +999,7 @@ G03:00:00 "GroundEpochName" @REQUEST_BEGIN("request2.name")
 
 `;
 
-    const actual = sequenceToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
+    const actual = seqnToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
     const expected = {
       id: 'id',
       metadata: {},
@@ -1204,7 +1203,7 @@ E-00:06:40.333 BAKE_BREAD`;
         },
       ],
     };
-    const actual = sequenceToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
+    const actual = seqnToSeqJson(SeqnParser.parse(seq), seq, commandDictionary, id);
     expect(actual).toEqual(expectedJson);
   });
 
@@ -1216,9 +1215,9 @@ E-00:06:40.333 BAKE_BREAD`;
   # comment
   C FSW_CMD_1 0.123 -2.34 # inline description`;
 
-      const seqJson1 = await sequenceToSeqJson(SeqnParser.parse(input), input, commandDictionary, 'id');
+      const seqJson1 = await seqnToSeqJson(SeqnParser.parse(input), input, commandDictionary, 'id');
       const seqN1 = await seqJsonToSequence(seqJson1);
-      const seqJson2 = await sequenceToSeqJson(SeqnParser.parse(seqN1), seqN1, commandDictionary, 'id');
+      const seqJson2 = await seqnToSeqJson(SeqnParser.parse(seqN1), seqN1, commandDictionary, 'id');
       expect(seqJson1).toEqual(seqJson2);
     });
 
@@ -1246,9 +1245,9 @@ G+3 "GroundEpochName" @REQUEST_BEGIN("request2.name")
 @METADATA "foo" "bar"
   `;
 
-      const seqJson1 = await sequenceToSeqJson(SeqnParser.parse(input), input, commandDictionary, 'id');
+      const seqJson1 = await seqnToSeqJson(SeqnParser.parse(input), input, commandDictionary, 'id');
       const seqN1 = await seqJsonToSequence(seqJson1);
-      const seqJson2 = await sequenceToSeqJson(SeqnParser.parse(seqN1), seqN1, commandDictionary, 'id');
+      const seqJson2 = await seqnToSeqJson(SeqnParser.parse(seqN1), seqN1, commandDictionary, 'id');
       expect(seqJson1).toEqual(seqJson2);
     });
   });
@@ -1261,7 +1260,7 @@ C CMD_0 true false [ false true ]
 @METADATA "foo" "bar"
 @MODEL "a" true "00:00:00"
   `;
-  const actual = sequenceToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
+  const actual = seqnToSeqJson(SeqnParser.parse(input), input, commandBanana, 'id');
   const expected = {
     id: 'id',
     metadata: {},
@@ -1328,7 +1327,7 @@ X_RANGE INT "...10" "-1,0,3,4,5,9"
   ];
 
   it('should return undefined if there are no variable children', () => {
-    let tree = parser.parse(`${seqN[0]}`);
+    let tree = SeqnParser.parse(`${seqN[0]}`);
     let variable = parseVariables(tree.topNode, seqN[0], 'LocalDeclaration');
     expect(variable).toBeDefined();
 
@@ -1352,7 +1351,7 @@ X_RANGE INT "...10" "-1,0,3,4,5,9"
       expect(variable[3]).toEqual({ name: 'CHARGE', type: 'INT' });
     }
 
-    tree = parser.parse(`${seqN[1]}`);
+    tree = SeqnParser.parse(`${seqN[1]}`);
     variable = parseVariables(tree.topNode, seqN[1], 'ParameterDeclaration');
     expect(variable).toBeDefined();
