+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './languages/seq-n/from-seq-json';
-export * from './languages/seq-n/to-seq-json';
+export * from './converters/seqJsonToSeqn';
+export * from './converters/seqnToSeqJson';
 export * from './languages/seq-n/seq-n';
